+++ cspackages-test/csharp-assembly-utils.lisp
@@ -1,6 +1,6 @@
 ;;; Generated automatically. Do not edit.
-;;; Generator Version: 31
-;;; Creation Date: 2026-07-04T16:50:34Z
+;;; Generator Version: 32
+;;; Creation Date: 2026-07-05T03:30:42Z
 
 ;;; Douglas P. Fields, Jr. - symbolics@lisp.engineer
 ;;; Copyright 2026 Douglas P. Fields, Jr.