+++ cspackages-test/system-net-service-point-manager.lisp
@@ -1,14 +1,14 @@
 ;;; Generated automatically. Do not edit.
 ;;; Class: System.Net.ServicePointManager
-;;; Generator Version: 31
-;;; Creation Date: 2026-07-04T16:50:34Z
+;;; Generator Version: 32
+;;; Creation Date: 2026-07-05T03:30:42Z
 
 (cl:in-package :system-net-service-point-manager)
 
 (cl:defconstant <type> (dotnet:resolve-type "System.Net.ServicePointManager"))
 (cl:defconstant <type-str> "System.Net.ServicePointManager")
-(cl:defconstant <creation> "2026-07-04T16:50:34Z")
-(cl:defconstant <version> 31)
+(cl:defconstant <creation> "2026-07-05T03:30:42Z")
+(cl:defconstant <version> 32)
 
 ;; Register C# Type with CLOS
 (cl:eval-when (:compile-toplevel :load-toplevel :execute)